requirement.txt
streamlit
streamlit-folium
pandas
numpy
geopandas
scikit-learn
matplotlib
seaborn
folium
psycopg2-binary
bcrypt
